requirements.txt
streamlit==1.38.0
google-generativeai==0.7.2
PyMuPDF==1.24.9
pdfplumber==0.11.4
numpy==1.26.4
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
